refactor: SoundTrackController로 통합
