+++ src/main/java/com/kl/grooveo/boundedContext/library/controller/SoundTrackController.java
@@ -32,6 +32,7 @@ import com.kl.grooveo.boundedContext.comment.entity.SoundPostComment;
 import com.kl.grooveo.boundedContext.comment.service.SoundPostCommentService;
 import com.kl.grooveo.boundedContext.library.dto.SoundTrackFormDTO;
 import com.kl.grooveo.boundedContext.library.entity.SoundTrack;
+import com.kl.grooveo.boundedContext.library.service.FileInfoService;
 import com.kl.grooveo.boundedContext.library.service.SoundTrackService;
 import com.kl.grooveo.boundedContext.reply.dto.ReplyFormDTO;
 
@@ -46,6 +47,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class SoundTrackController {
 	private final SoundTrackService soundTrackService;
+	private final FileInfoService fileInfoService;
 	private final SoundPostCommentService soundPostCommentService;
 	private final Rq rq;
 	private final AmazonS3 amazonS3Client;
@@ -71,6 +73,69 @@ public class SoundTrackController {
 		model.addAttribute("soundTracks", soundTracks);
 
 		return "usr/library/soundTrackList";
+	}
+
+	@PreAuthorize("isAuthenticated()")
+	@GetMapping("/soundupload")
+	public String showSoundUpload(Model model, SoundTrackFormDTO soundTrackForm) {
+		return "usr/library/soundUpload";
+	}
+
+	@PreAuthorize("isAuthenticated()")
+	@PostMapping("/soundupload")
+	public String uploadFiles(Model model, @Valid SoundTrackFormDTO soundTrackFormDTO, BindingResult bindingResult) {
+		try {
+			if (soundTrackFormDTO.getAlbumCover().isEmpty() || soundTrackFormDTO.getSoundFile().isEmpty()) {
+				bindingResult.rejectValue("file", "required", "음원과 앨범 등록은 필수입니다.");
+				return "redirect:/library/soundUpload";
+			}
+
+			String albumCoverExtension = FilenameUtils.getExtension(
+				soundTrackFormDTO.getAlbumCover().getOriginalFilename());
+			String albumCoverName = UUID.randomUUID().toString() + "." + albumCoverExtension;
+			String albumCoverUrl =
+				"https://s3." + region + ".amazonaws.com/" + bucket + "/albumCover/" + albumCoverName;
+
+			String soundExtension = FilenameUtils.getExtension(soundTrackFormDTO.getSoundFile().getOriginalFilename());
+			String soundName = UUID.randomUUID().toString() + "." + soundExtension;
+			String soundUrl = "https://s3." + region + ".amazonaws.com/" + bucket + "/sound/" + soundName;
+
+			String title = URLEncoder.encode(soundTrackFormDTO.getTitle(), StandardCharsets.UTF_8);
+			String description = URLEncoder.encode(soundTrackFormDTO.getDescription(), StandardCharsets.UTF_8);
+
+			ObjectMetadata albumCoverMetadata = new ObjectMetadata();
+			albumCoverMetadata.setContentType(soundTrackFormDTO.getAlbumCover().getContentType());
+			albumCoverMetadata.setContentLength(soundTrackFormDTO.getAlbumCover().getSize());
+			albumCoverMetadata.addUserMetadata("title", title);
+			albumCoverMetadata.addUserMetadata("description", description);
+
+			ObjectMetadata soundMetadata = new ObjectMetadata();
+			soundMetadata.setContentType(soundTrackFormDTO.getSoundFile().getContentType());
+			soundMetadata.setContentLength(soundTrackFormDTO.getSoundFile().getSize());
+			soundMetadata.addUserMetadata("title", title);
+			soundMetadata.addUserMetadata("description", description);
+
+			amazonS3Client.putObject(new PutObjectRequest(bucket, "albumCover/" + albumCoverName,
+				soundTrackFormDTO.getAlbumCover().getInputStream(), albumCoverMetadata));
+			amazonS3Client.putObject(
+				new PutObjectRequest(bucket, "sound/" + soundName, soundTrackFormDTO.getSoundFile().getInputStream(),
+					soundMetadata));
+
+			SoundTrack fileInfo = SoundTrack.builder()
+				.title(soundTrackFormDTO.getTitle())
+				.artist(rq.getMember())
+				.description(soundTrackFormDTO.getDescription())
+				.albumCoverUrl(albumCoverUrl)
+				.soundUrl(soundUrl)
+				.build();
+
+			fileInfoService.saveFileInfo(fileInfo);
+
+			return "redirect:/library/soundDetail/" + fileInfo.getId();
+		} catch (IOException e) {
+			e.printStackTrace();
+			return "redirect:/library/soundUpload";
+		}
 	}
 
 	@PreAuthorize("isAuthenticated()")
